refactor(brett): extract presets.ts (SPEC, validateAppearance, preset CRUD)
